AgroVision-ML-API/api/requirements.txt
flask
numpy
pillow
scikit-image
firebase-admin
tensorflow